Text.Pandoc.Options

Description

Data structures and functions for representing parser and writer options.

Documentation

data Extension

Individually selectable syntax extensions.

Constructors

Ext_footnotes
PandocPHPMMD style footnotes

Ext_inline_notes
Pandoc-style inline notes

Ext_pandoc_title_block
Pandoc title block

Ext_yaml_metadata_block
YAML metadata block

Ext_mmd_title_block
Multimarkdown metadata block

Ext_table_captions
Pandoc-style table captions

Ext_implicit_figures
A paragraph with just an image is a figure

Ext_simple_tables
Pandoc-style simple tables

Ext_multiline_tables
Pandoc-style multiline tables

Ext_grid_tables
Grid tables (pandoc, reST)

Ext_pipe_tables
Pipe tables (as in PHP markdown extra)

Ext_citations
Pandoc/citeproc citations

Ext_raw_tex
Allow raw TeX (other than math)

Ext_raw_html
Allow raw HTML

Ext_tex_math_dollars
TeX math between $..$ or $$..$$

Ext_tex_math_single_backslash
TeX math btw (..) [..]

Ext_tex_math_double_backslash
TeX math btw \(..\) \[..\]

Ext_latex_macros
Parse LaTeX macro definitions (for math only)

Ext_fenced_code_blocks
Parse fenced code blocks

Ext_fenced_code_attributes
Allow attributes on fenced code blocks

Ext_backtick_code_blocks
Github style ``` code blocks

Ext_inline_code_attributes
Allow attributes on inline code

Ext_markdown_in_html_blocks
Interpret as markdown inside HTML blocks

Ext_markdown_attribute
Interpret text inside HTML as markdown iff container has attribute markdown

Ext_escaped_line_breaks
Treat a backslash at EOL as linebreak

Ext_link_attributes
MMD style reference link attributes

Ext_autolink_bare_uris
Make all absolute URIs into links

Ext_fancy_lists
Enable fancy list numbers and delimiters

Ext_lists_without_preceding_blankline
Allow lists without preceding blank

Ext_startnum
Make start number of ordered list significant

Ext_definition_lists
Definition lists as in pandoc, mmd, php

Ext_example_lists
Markdown-style numbered examples

Ext_all_symbols_escapable
Make all non-alphanumerics escapable

Ext_intraword_underscores
Treat underscore inside word as literal

Ext_blank_before_blockquote
Require blank line before a blockquote

Ext_blank_before_header
Require blank line before a header

Ext_strikeout
Strikeout using ~~this~~ syntax

Ext_superscript
Superscript using ^this^ syntax

Ext_subscript
Subscript using ~this~ syntax

Ext_hard_line_breaks
All newlines become hard line breaks

Ext_ignore_line_breaks
Newlines in paragraphs are ignored

Ext_literate_haskell
Enable literate Haskell conventions

Ext_abbreviations
PHP markdown extra abbreviation definitions

Ext_auto_identifiers
Automatic identifiers for headers

Ext_ascii_identifiers
ascii-only identifiers for headers

Ext_header_attributes
Explicit header attributes {#id .class k=v}

Ext_mmd_header_identifiers
Multimarkdown style header identifiers [myid]

Ext_implicit_header_references
Implicit reference links for headers

Ext_line_blocks
RST style line blocks

pandocExtensions :: Set Extension

strictExtensions :: Set Extension

phpMarkdownExtraExtensions :: Set Extension

githubMarkdownExtensions :: Set Extension

multimarkdownExtensions :: Set Extension

data ReaderOptions

Constructors

ReaderOptions

Fields

readerExtensions :: Set Extension

Syntax extensions

readerSmart :: Bool

Smart punctuation

readerStrict :: Bool

FOR TRANSITION ONLY

readerStandalone :: Bool

Standalone document with header

readerParseRaw :: Bool

Parse raw HTML, LaTeX

readerColumns :: Int

Number of columns in terminal

readerTabStop :: Int

Tab stop

readerOldDashes :: Bool

Use pandoc <= 1.8.2.1 behavior in parsing dashes; -- is em-dash; - before numerial is en-dash

readerApplyMacros :: Bool

Apply macros to TeX math

readerIndentedCodeClasses :: [String]

Default classes for indented code blocks

readerDefaultImageExtension :: String

Default extension for images

readerTrace :: Bool

Print debugging info

data HTMLMathMethod

Constructors

PlainMath

LaTeXMathML (Maybe String)

JsMath (Maybe String)

GladTeX

WebTeX String

MathML (Maybe String)

MathJax String

data CiteMethod

Constructors

Citeproc

Natbib

Biblatex

data ObfuscationMethod

Methods for obfuscating email addresses in HTML.

Constructors

NoObfuscation

ReferenceObfuscation

JavascriptObfuscation

data HTMLSlideVariant

Varieties of HTML slide shows.

Constructors

S5Slides

SlidySlides

SlideousSlides

DZSlides

RevealJsSlides

NoSlides

data EPUBVersion

Constructors

EPUB2

EPUB3

data WriterOptions

Options for writers

Constructors

WriterOptions

Fields

writerStandalone :: Bool

Include header and footer

writerTemplate :: String

Template to use in standalone mode

writerVariables :: [(String, String)]

Variables to set in template

writerTabStop :: Int

Tabstop for conversion btw spaces and tabs

writerTableOfContents :: Bool

Include table of contents

writerSlideVariant :: HTMLSlideVariant

Are we writing S5, Slidy or Slideous?

writerIncremental :: Bool

True if lists should be incremental

writerHTMLMathMethod :: HTMLMathMethod

How to print math in HTML

writerIgnoreNotes :: Bool

Ignore footnotes (used in making toc)

writerNumberSections :: Bool

Number sections in LaTeX

writerNumberOffset :: [Int]

Starting number for section, subsection, ...

writerSectionDivs :: Bool

Put sections in div tags in HTML

writerExtensions :: Set Extension

Markdown extensions that can be used

writerReferenceLinks :: Bool

Use reference links in writing markdown, rst

writerWrapText :: Bool

Wrap text to line length

writerColumns :: Int

Characters in a line (for text wrapping)

writerEmailObfuscation :: ObfuscationMethod

How to obfuscate emails

writerIdentifierPrefix :: String

Prefix for section & note ids in HTML and for footnote marks in markdown

writerSourceURL :: Maybe String

Absolute URL + directory of 1st source file

writerUserDataDir :: Maybe FilePath

Path of user data directory

writerCiteMethod :: CiteMethod

How to print cites

writerHtml5 :: Bool

Produce HTML5

writerHtmlQTags :: Bool

Use q tags for quotes in HTML

writerBeamer :: Bool

Produce beamer LaTeX slide show

writerSlideLevel :: Maybe Int

Force header level of slides

writerChapters :: Bool

Use "chapter" for top-level sects

writerListings :: Bool

Use listings package for code

writerHighlight :: Bool

Highlight source code

writerHighlightStyle :: Style

Style to use for highlighting

writerSetextHeaders :: Bool

Use setext headers for levels 1-2 in markdown

writerTeXLigatures :: Bool

Use tex ligatures quotes, dashes in latex

writerEpubVersion :: Maybe EPUBVersion

Nothing or EPUB version

writerEpubMetadata :: String

Metadata to include in EPUB

writerEpubStylesheet :: Maybe String

EPUB stylesheet specified at command line

writerEpubFonts :: [FilePath]

Paths to fonts to embed

writerEpubChapterLevel :: Int

Header level for chapters (separate files)

writerTOCDepth :: Int

Number of levels to include in TOC

writerReferenceODT :: Maybe FilePath

Path to reference ODT if specified

writerReferenceDocx :: Maybe FilePath

Ptah to reference DOCX if specified

def :: Default a => a

isEnabled :: Extension -> WriterOptions -> Bool

Returns True if the given extension is enabled.