Text.Pandoc.Readers.TeXMath

Description

Conversion of TeX math to a list of Pandoc inline elements.

Documentation

readTeXMath

Arguments

:: String
String to parse (assumes '\n' line endings)

-> [Inline]

Deprecated: Use readTeXMath' from Text.Pandoc.JSON instead

Converts a raw TeX math formula to a list of Pandoc inlines. Defaults to raw formula between $ characters if entire formula can't be converted. (This is provided for backwards compatibility; it is better to use readTeXMath', which properly distinguishes between display and inline math.)

readTeXMath'

Arguments

:: MathType

-> String
String to parse (assumes '\n' line endings)

-> [Inline]

Converts a raw TeX math formula to a list of Pandoc inlines. Defaults to raw formula between $ or $$ characters if entire formula can't be converted.