Text.Pandoc.Readers.Txt2Tags

Description

Conversion of txt2tags formatted plain text to Pandoc document.

Documentation

readTxt2Tags :: T2TMeta -> ReaderOptions -> String -> Pandoc

Read Txt2Tags from an input string returning a Pandoc document

getT2TMeta :: [FilePath] -> FilePath -> IO T2TMeta

Get the meta information required by Txt2Tags macros

data T2TMeta

An object for the T2T macros meta information the contents of each field is simply substituted verbatim into the file

Constructors

T2TMeta

Fields

date :: String

Current date

mtime :: String

Last modification time of infile

infile :: FilePath

Input file

outfile :: FilePath

Output file

readTxt2TagsNoMacros :: ReaderOptions -> String -> Pandoc

Read Txt2Tags (ignoring all macros) from an input string returning a Pandoc document