Text.Pandoc.SelfContained

Description

Functions for converting an HTML file into one that can be viewed offline, by incorporating linked images, CSS, and scripts into the HTML using data URIs.

Documentation

makeSelfContained :: Maybe FilePath -> String -> IO String

Convert HTML into self-contained HTML, incorporating images, scripts, and CSS using data: URIs. Items specified using absolute URLs will be downloaded; those specified using relative URLs will be sought first relative to the working directory, then relative to the user data directory (if the first parameter is Just a directory), and finally relative to pandoc's default data directory.