Text.Pandoc.Shared

Description

Utility functions and definitions used by the various Pandoc modules.

List processing

splitBy :: (a -> Bool) -> [a] -> [[a]]

Split list by groups of one or more sep.

splitByIndices :: [Int] -> [a] -> [[a]]

splitStringByIndices :: [Int] -> [Char] -> [[Char]]

Split string into chunks divided at specified indices.

substitute :: Eq a => [a] -> [a] -> [a] -> [a]

Replace each occurrence of one sublist in a list with another.

Text processing

backslashEscapes

Arguments

:: [Char]
list of special characters to escape

-> [(Char, String)]

Returns an association list of backslash escapes for the designated characters.

escapeStringUsing :: [(Char, String)] -> String -> String

Escape a string of characters, using an association list of characters and strings.

stripTrailingNewlines :: String -> String

Strip trailing newlines from string.

trim :: String -> String

Remove leading and trailing space (including newlines) from string.

triml :: String -> String

Remove leading space (including newlines) from string.

trimr :: String -> String

Remove trailing space (including newlines) from string.

stripFirstAndLast :: String -> String

Strip leading and trailing characters from string

camelCaseToHyphenated :: String -> String

Change CamelCase word to hyphenated lowercase (e.g., camel-case).

toRomanNumeral :: Int -> String

Convert number < 4000 to uppercase roman numeral.

escapeURI :: String -> String

Escape whitespace in URI.

tabFilter

Arguments

:: Int
Tab stop

-> String
Input

-> String

Convert tabs to spaces and filter out DOS line endings. Tabs will be preserved if tab stop is set to 0.

Date/time

normalizeDate :: String -> Maybe String

Parse a date and convert (if possible) to YYYY-MM-DD format.

Pandoc block and inline list processing

orderedListMarkers :: (Int, ListNumberStyle, ListNumberDelim) -> [String]

Generate infinite lazy list of markers for an ordered list, depending on list attributes.

normalizeSpaces :: [Inline] -> [Inline]

Normalize a list of inline elements: remove leading and trailing Space elements, collapse double Spaces into singles, and remove empty Str elements.

normalize :: (Eq a, Data a) => a -> a

Normalize Pandoc document, consolidating doubled Spaces, combining adjacent Strs and Emphs, remove Nulls and empty elements, etc.

stringify :: Walkable Inline a => a -> String

Convert pandoc structure to a string with formatting removed. Footnotes are skipped (since we don't want their contents in link labels).

compactify

Arguments

:: [[Block]]
List of list items (each a list of blocks)

-> [[Block]]

Change final list item from Para to Plain if the list contains no other Para blocks.

compactify'

Arguments

:: [Blocks]
List of list items (each a list of blocks)

-> [Blocks]

Change final list item from Para to Plain if the list contains no other Para blocks. Like compactify, but operates on Blocks rather than [Block].

compactify'DL :: [(Inlines, [Blocks])] -> [(Inlines, [Blocks])]

Like compactify', but akts on items of definition lists.

data Element

Data structure for defining hierarchical Pandoc documents

Constructors

Blk Block

Sec Int [Int] Attr [Inline] [Element]

hierarchicalize :: [Block] -> [Element]

Convert list of Pandoc blocks into (hierarchical) list of Elements

uniqueIdent :: [Inline] -> [String] -> String

Generate a unique identifier from a list of inlines. Second argument is a list of already used identifiers.

isHeaderBlock :: Block -> Bool

True if block is a Header block.

headerShift :: Int -> Pandoc -> Pandoc

Shift header levels up or down.

isTightList :: [[Block]] -> Bool

Detect if a list is tight.

addMetaField :: ToMetaValue a => String -> a -> Meta -> Meta

Set a field of a Meta object. If the field already has a value, convert it into a list with the new value appended to the old value(s).

makeMeta :: [Inline] -> [[Inline]] -> [Inline] -> Meta

Create Meta from old-style title, authors, date. This is provided to ease the transition from the old API.

TagSoup HTML handling

renderTags' :: [Tag String] -> String

Render HTML tags.

File handling

inDirectory :: FilePath -> IO a -> IO a

Perform an IO action in a directory, returning to starting directory.

readDataFile :: Maybe FilePath -> FilePath -> IO ByteString

Read file from specified user data directory or, if not found there, from Cabal data directory.

readDataFileUTF8 :: Maybe FilePath -> FilePath -> IO String

Same as readDataFile but returns a String instead of a ByteString.

fetchItem :: Maybe String -> String -> IO (Either SomeException (ByteString, Maybe String))

Fetch an image or other item from the local filesystem or the net. Returns raw content and maybe mime type.

openURL :: String -> IO (Either SomeException (ByteString, Maybe String))

Read from a URL and return raw data and maybe mime type.

Error handling

err :: Int -> String -> IO a

warn :: String -> IO ()

Safe read

safeRead :: (Monad m, Read a) => String -> m a