Text.Pandoc.Writers.Native

Description

Conversion of a Pandoc document to a string representation.

Note: If writerStandalone is False, only the document body is represented; otherwise, the full Pandoc document, including the metadata.

Documentation

writeNative :: WriterOptions -> Pandoc -> String

Prettyprint Pandoc document.