Text.Pandoc.XML

Description

Functions for escaping and formatting XML.

Documentation

escapeCharForXML :: Char -> String

Escape one character as needed for XML.

escapeStringForXML :: String -> String

Escape string as needed for XML. Entity references are not preserved.

inTags :: Bool -> String -> [(String, String)] -> Doc -> Doc

Put the supplied contents between start and end tags of tagType, with specified attributes and (if specified) indentation.

selfClosingTag :: String -> [(String, String)] -> Doc

Return a self-closing tag of tagType with specified attributes

inTagsSimple :: String -> Doc -> Doc

Put the supplied contents between start and end tags of tagType.

inTagsIndented :: String -> Doc -> Doc

Put the supplied contents in indented block btw start and end tags.

toEntities :: String -> String

Escape all non-ascii characters using numerical entities.

fromEntities :: String -> String